CreateTmPiDataset

Documentation for CreateTmPiDataset.

CreateTmPiDataset.cdskey

CreateTmPiDataset.parserequest

CreateTmPiDataset.retrieve

CreateTmPiDataset.cdskey — Method

cdskey() -> Dict{Any,Any}

Retrieves the CDS API ckeys from the ~/.cdsapirc file in the home directory

CreateTmPiDataset.parserequest — Method

parserequest(
    ckeys :: AbstractDict,
    resp  :: AbstractDict,
    api   :: AbstractString
) -> Dict{Any,Any}

Get info on HTTP request, and parse the information and update the dictionary

Arguments

ckeys : Dictionary that contains the CDSAPI information held in ~/.cdsapirc

resp : Dictionary that contains the HTTP response

api : String that contains the API used for Climate Data Store authentication

CreateTmPiDataset.retrieve — Function

retrieve(
    dset  :: AbstractString,
    dkeys :: AbstractDict,
    fnc   :: AbstractString,
    ckeys :: AbstractDict = cdskey()
)

Retrieves dsets from the Climate Data Store, with options specified in a Julia Dictionary and saves it into a specified file. Arguments: * dset::AbstractString : string specifies the name of the dset within the Climate Data Store that the retrieve function is attempting to retrieve data from * dkeys::AbstractDict : dictionary that contains the keywords that specify the properties (e.g. date, resolution, grid) of the data being retrieved * fnc::AbstractString : string that contains the path and name of the file that the data is to be saved into * ckeys::AbstractDict : dictionary that contains API Key information read from the .cdsapirc file in the home directory (optional)